No longer define anything about the storage type
I had an important breakthrough, in which I realized that string
and string_view implement the same "concept" of what a "String" is.

The encoding types therefore become a simple wrapper around a
"String-concept" type at the same time that it offers a
"String-concept"-compatible type as well.

The only thing that we do is a few static asserts to make sure that
the underlying storage is actually compatible with the expected
encoding.